added last created installer, fixed hex-output, refactored CMI
for demo purposes, I added the latest created installer to the source repo.  On my Windows it works well.
Inside, you find a fix to the hex output - missing commata added.
Refactoring of the CMake Integration ... well, I found CMakeIntegration to be a coding topic, also.  Especially for a Tool like an Assembler.  I read up more on CMake packages and found the minimum neccessary to have is a PackageConfig.cmake file.  The Rest will be generated and should be included from the PackageConfig.cmake.  Except for the PackageConfigVersion.cmake file - which gets generated and is loaded prior to the project itself for version checking (on installed system).
Anyhow... so building the CMake Integration seemed to me like a custom build step, so I implemented it as such.  ALL depends on it, so it's always built before installation or packaging.
Oh, and the package seems to have taken up Wanja's Icon :)